[#49] Refactor js extension to jsx
