+++ build-utils/webpack.common.js
@@ -14,11 +14,14 @@ const config = {
   module: {
     rules: [
       {
-        test: /\.(js)$/,
+        test: /\.(js|jsx)$/,
         exclude: /node_modules/,
         use: ["babel-loader"],
       },
     ],
+  },
+  resolve: {
+    extensions: [".js", ".jsx"],
   },
   optimization: {
     splitChunks: {
