+++ build-utils/webpack.dev.js
@@ -7,7 +7,7 @@ const port = process.env.PORT || 3000;
 const config = {
   mode: "development",
   entry: {
-    app: `${commonPaths.appEntry}/index.js`,
+    app: `${commonPaths.appEntry}/index.jsx`,
   },
   output: {
     filename: "[name].[hash].js",
