+++ build-utils/webpack.prod.js
@@ -6,7 +6,7 @@ const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 const config = {
   mode: "production",
   entry: {
-    app: [`${commonPaths.appEntry}/index.js`],
+    app: [`${commonPaths.appEntry}/index.jsx`],
   },
   output: {
     filename: "static/[name].[hash].js",
