+++ package.json
@@ -2,7 +2,7 @@
   "name": "react-starter-boilerplate-hmr",
   "version": "1.0.0",
   "description": "SPipes editor UI",
-  "main": "index.js",
+  "main": "index.jsx",
   "author": "Petr Jordan",
   "license": "MIT",
   "scripts": {
