+++ src/components/App.jsx
@@ -2,22 +2,22 @@ import React, { Component } from "react";
 import { Switch, BrowserRouter as Router, Route } from "react-router-dom";
 import importedComponent from "react-imported-component";
 
-import Home from "./Home";
-import Loading from "./Loading";
+import Home from "./Home.jsx";
+import Loading from "./Loading.jsx";
 
-const Scripts = importedComponent(() => import(/* webpackChunkName:'Scripts' */ "./Scripts"), {
+const Scripts = importedComponent(() => import(/* webpackChunkName:'Scripts' */ "./Scripts.jsx"), {
   LoadingComponent: Loading,
 });
-const Executions = importedComponent(() => import(/* webpackChunkName:'Executions' */ "./Executions"), {
+const Executions = importedComponent(() => import(/* webpackChunkName:'Executions' */ "./Executions.jsx"), {
   LoadingComponent: Loading,
 });
-const AsyncDynamicPAge = importedComponent(() => import(/* webpackChunkName:'DynamicPage' */ "./DynamicPage"), {
+const AsyncDynamicPAge = importedComponent(() => import(/* webpackChunkName:'DynamicPage' */ "./DynamicPage.jsx"), {
   LoadingComponent: Loading,
 });
-const AsyncDagre = importedComponent(() => import(/* webpackChunkName:'Dagre' */ "./dagre/Dagre.js"), {
+const AsyncDagre = importedComponent(() => import(/* webpackChunkName:'Dagre' */ "./dagre/Dagre.jsx"), {
   LoadingComponent: Loading,
 });
-const AsyncNoMatch = importedComponent(() => import(/* webpackChunkName:'NoMatch' */ "./NoMatch"), {
+const AsyncNoMatch = importedComponent(() => import(/* webpackChunkName:'NoMatch' */ "./NoMatch.jsx"), {
   LoadingComponent: Loading,
 });
