+++ src/components/NoMatch.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Icon } from "semantic-ui-react";
 
-import Layout from "./Layout";
+import Layout from "./Layout.jsx";
 
 const NoMatch = () => {
   return (
