+++ src/components/ScriptFunctionSelection.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Dropdown } from "semantic-ui-react";
-import { FUNCTION, FUNCTION_NAME, Rest } from "./rest/Rest";
-import { ICONS_MAP } from "./dagre/DagreIcons";
+import { FUNCTION, FUNCTION_NAME, Rest } from "./rest/Rest.jsx";
+import { ICONS_MAP } from "./dagre/DagreIcons.jsx";
 
 class ScriptFunctionSelection extends React.Component {
   constructor(props) {
