+++ src/components/modal/ScriptOntologyModal.jsx
@@ -14,12 +14,12 @@ import {
   SCRIPT_PATH,
   START_DATE_UNIX,
   TRANSFORMATION,
-} from "../rest/Rest";
+} from "../rest/Rest.jsx";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faEdit, faMugHot, faPlayCircle, faPlusCircle, faTrash } from "@fortawesome/free-solid-svg-icons";
 import Moment from "react-moment";
 import { Link } from "react-router-dom";
-import Layout from "../Layout";
+import Layout from "../Layout.jsx";
 
 class ScriptOntologyModal extends React.Component {
   constructor(props) {
