+++ src/components/sform/SFormsFunctionModal.jsx
@@ -1,48 +1,45 @@
 import React from "react";
 import SForms from "@kbss-cvut/s-forms";
-import { Button, Modal } from "react-bootstrap";
-import { Rest } from "../rest/Rest";
-import Loading from "../Loading";
-import "@triply/yasgui/build/yasgui.min.css";
-import ErrorModal from "../modal/ErrorModal";
 
-class SFormsModal extends React.Component {
+import { Button, Modal } from "react-bootstrap";
+import { Rest } from "../rest/Rest.jsx";
+import "@triply/yasgui/build/yasgui.min.css";
+import ErrorModal from "../modal/ErrorModal.jsx";
+import ScriptActionsModuleModal from "../modal/ScriptActionsModuleModal.jsx";
+
+class SFormsFunctionModal extends React.Component {
   constructor(props) {
     super(props);
 
     this.state = {
       isLoaded: false,
-      isLoading: false,
       modalVisible: false,
       selectedForm: null,
       moduleTypeUri: null,
       moduleUri: null,
-      errorMessage: null,
       scriptPath: null,
+      errorMessage: null,
     };
     this.refForm = React.createRef();
     this.handleErrorModal = this.handleErrorModal.bind(this);
   }
 
   componentWillReceiveProps(newProps) {
-    if (newProps.moduleTypeUri && newProps.scriptPath) {
-      this.setState({ isLoading: true });
-      Rest.getScriptForm(newProps.moduleTypeUri, newProps.moduleUri, newProps.scriptPath).then((response) => {
+    if (newProps.scriptPath && newProps.functionUri) {
+      Rest.getFunctionForm(newProps.scriptPath, newProps.functionUri).then((response) => {
         this.setState({
           isLoaded: true,
-          isLoading: false,
           selectedForm: response,
           modalVisible: true,
-          moduleTypeUri: newProps.moduleTypeUri,
-          moduleUri: newProps.moduleUri,
-          scriptPath: newProps.scriptPath,
+          moduleTypeUri: newProps.scriptPath,
+          moduleUri: newProps.functionUri,
         });
       });
     }
   }
 
   handleClose() {
-    this.setState({ modalVisible: false, isLoaded: false, isLoading: false });
+    this.setState({ modalVisible: false });
   }
 
   handleErrorModal() {
@@ -50,19 +47,36 @@ class SFormsModal extends React.Component {
   }
 
   handleSubmit() {
-    let form = this.state.selectedForm;
-    form["http://onto.fel.cvut.cz/ontologies/documentation/has_related_question"] =
-      this.refForm.current.context.getFormQuestionsData();
+    let data =
+      this.refForm.current.context.getFormQuestionsData()[0][
+        "http://onto.fel.cvut.cz/ontologies/documentation/has_related_question"
+      ];
+    let functionUri = "";
+    let params = [];
+    data.forEach((q) => {
+      let label = q["http://www.w3.org/2000/01/rdf-schema#label"] + "=";
+      let value =
+        q["http://onto.fel.cvut.cz/ontologies/documentation/has_answer"][0][
+          "http://onto.fel.cvut.cz/ontologies/documentation/has_data_value"
+        ];
+      if (value["@value"] !== undefined) {
+        params.push(label + value["@value"]);
+      } else if (label === "URI=") {
+        functionUri = value;
+      }
+    });
 
-    this.setState({ isLoading: true, isLoaded: false });
-    Rest.updateScriptForm(this.state.moduleTypeUri, form, this.state.scriptPath).then((response) => {
-      this.setState({ isLoading: false });
+    Rest.executeFunction(functionUri, params.join("&")).then((response) => {
+      console.log(response);
+      console.log(response.status);
       if (response.status === 200) {
-        window.location.reload(false);
+        console.log(response);
+        window.location.href = "/executions";
       } else {
-        console.log("ERROR on script update");
+        console.log("ERROR during script execution");
         this.setState({ errorMessage: "ERROR during script execution" });
       }
+      this.setState({ isLoaded: false, modalVisible: false });
     });
   }
 
@@ -86,8 +100,6 @@ class SFormsModal extends React.Component {
 
     if (this.state.errorMessage) {
       return <ErrorModal errorMessage={this.state.errorMessage} handleErrorModal={this.handleErrorModal} />;
-    } else if (this.state.isLoading) {
-      return <Loading size={"large"} style={{ margin: "auto", position: "absolute", inset: "0px", zIndex: 9000 }} />;
     } else if (this.state.isLoaded) {
       return (
         <Modal
@@ -118,4 +130,4 @@ class SFormsModal extends React.Component {
   }
 }
 
-export default SFormsModal;
+export default SFormsFunctionModal;
