+++ src/components/sform/ScriptInputOutputModal.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { Alert, Button, Form, Modal } from "react-bootstrap";
-import { Rest } from "../rest/Rest";
+import { Rest } from "../rest/Rest.jsx";
 import Yasgui from "@triply/yasgui";
 import "@triply/yasgui/build/yasgui.min.css";
 import JSONPretty from "react-json-pretty";
