--- src/index.js
@@ -1,10 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./components/App";
-
-import "./index.css";
-import "bootstrap/dist/css/bootstrap.css";
-
-const render = (Component) => ReactDOM.render(<Component />, document.getElementById("root"));
-
-render(App);
