+++ src/index.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+
+import App from "./components/App.jsx";
+import "./index.css";
+import "bootstrap/dist/css/bootstrap.css";
+
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(<App />);
